fenrir_vk_2.76
fenrir_vk_2.76

+++ ealvatag/src/main/java/ealvatag/audio/Utils.java
@@ -83,7 +83,7 @@ public class Utils {
     public static long getLongLE(ByteBuffer b, int start, int end) {
         long number = 0;
         for (int i = 0; i < (end - start + 1); i++) {
-            number += ((b.get(start + i) & 0xFF) << i * 8);
+            number += ((long) (b.get(start + i) & 0xFF) << i * 8);
         }
 
         return number;
